adding character count and spliting in diferent files
