+++ holberton.h
@@ -20,5 +20,8 @@ void (*f)();
 int _putchar(char c);
 int _printf(const char *format, ...);
 void print_number(int n);
+void print_char(va_list c);
+void print_string(va_list s);
+void print_percentage(void);
 
 #endif
